List of books done
Missing: Description, edit books function and add a book funktion.

+++ UtilLibrary/MsSqlRepsoitory/Model/Items/IBooks.cs
@@ -5,7 +5,7 @@ namespace UtilLibrary.MsSqlRepsoitory
     public interface IBooks
     {
         string Author { get; set; }
-        int ISBN { get; set; }
+        Int64 ISBN { get; set; }
         string Category { get; set; }
         int ItemsID { get; set; }
         int Pages { get; set; }
