+++ UtilLibrary/MsSqlRepsoitory/Model/Items/IFullBooks.cs
@@ -5,7 +5,7 @@ namespace UtilLibrary.MsSqlRepsoitory
     public interface IFullBooks : IItems
     {
         string Author { get; set; }
-        int ISBN { get; set; }
+        long ISBN { get; set; }
         string Category { get; set; }
         int Pages { get; set; }
         string Publisher { get; set; }
